Fix validating webhook
The common label `part-of` is used by the kubernetesDeployment (which is the web server for the validating webhooks). The value is `kubernetes` there and the service selector uses it. Now we overwrite the label in the kustomization file and end up selecting the apiDeployment and kubernetesDeployemnt as endpoints for the kubernetesService. This breaks every other `kubectl apply` of a pyrra CRD.